+++ janitor/functions.py
@@ -6,7 +6,7 @@ import unicodedata
 import warnings
 from fnmatch import translate
 from functools import partial, reduce
-from typing import Any, Callable, Dict, Iterable, List, Tuple, Union
+from typing import Any, Callable, Dict, Iterable, List, Tuple, Union, Set
 
 import numpy as np
 import pandas as pd
@@ -3485,3 +3485,39 @@ def count_cumulative_unique(
     )
 
     return df
+
+
+@pf.register_series_method
+def toset(series: pd.Series) -> Set:
+    """
+    Return a set of the values.
+
+    These are each a scalar type, which is a Python scalar
+    (for str, int, float) or a pandas scalar
+    (for Timestamp/Timedelta/Interval/Period)
+
+    Functional usage example:
+
+    .. code-block:: python
+
+        import pandas as pd
+        import janitor as jn
+
+        series = pd.Series(...)
+        s = jn.functions.toset(series=series)
+
+    Method chaining usage example:
+
+    .. code-block:: python
+
+        import pandas as pd
+        import janitor.functions
+
+        series = pd.Series(...)
+        s = series.toset()
+
+    :param series: A pandas series.
+    :returns: A set of values.
+    """
+
+    return set(series.tolist())